@@ -15,19 +15,19 @@
   },
   "repository": "https://github.com/LemmyNet/lemmy-ui",
   "dependencies": {
-    "@typescript-eslint/parser": "^4.14.2",
+    "@typescript-eslint/parser": "^4.15.1",
     "autosize": "^4.0.2",
     "choices.js": "^9.0.1",
     "emoji-short-name": "^1.0.0",
     "express": "~4.17.1",
-    "i18next": "^19.8.7",
-    "inferno": "^7.4.7",
-    "inferno-create-element": "^7.4.7",
+    "i18next": "^19.8.9",
+    "inferno": "^7.4.8",
+    "inferno-create-element": "^7.4.8",
     "inferno-helmet": "^5.2.1",
-    "inferno-hydrate": "^7.4.7",
+    "inferno-hydrate": "^7.4.8",
     "inferno-i18next": "github:nimbusec-oss/inferno-i18next#semver:^7.4.2",
-    "inferno-router": "^7.4.7",
-    "inferno-server": "^7.4.7",
+    "inferno-router": "^7.4.8",
+    "inferno-server": "^7.4.8",
     "isomorphic-cookie": "^1.2.4",
     "jwt-decode": "^3.1.2",
     "markdown-it": "^12.0.4",
@@ -45,17 +45,18 @@
     "ws": "^7.4.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.13",
-    "@babel/plugin-transform-runtime": "^7.12.15",
-    "@babel/plugin-transform-typescript": "^7.12.13",
-    "@babel/preset-env": "7.12.13",
-    "@babel/preset-typescript": "^7.12.13",
-    "@babel/runtime": "^7.12.13",
+    "@babel/core": "^7.12.17",
+    "@babel/plugin-transform-runtime": "^7.12.17",
+    "@babel/plugin-transform-typescript": "^7.12.17",
+    "@babel/preset-env": "7.12.17",
+    "@babel/preset-typescript": "^7.12.17",
+    "@babel/runtime": "^7.12.18",
     "@types/autosize": "^3.0.6",
     "@types/express": "^4.17.11",
-    "@types/node": "^14.14.25",
+    "@types/node": "^14.14.31",
     "@types/node-fetch": "^2.5.8",
     "@types/serialize-javascript": "^5.0.0",
+    "@typescript-eslint/eslint-plugin": "^4.15.1",
     "babel-loader": "^8.2.2",
     "babel-plugin-inferno": "^6",
     "bootstrap": "^4.6.0",
@@ -63,23 +64,23 @@
     "clean-webpack-plugin": "^3.0.0",
     "copy-webpack-plugin": "^7.0.0",
     "css-loader": "^5.0.2",
-    "eslint": "^7.19.0",
-    "eslint-plugin-jane": "^9.0.6",
-    "husky": "^4.3.8",
+    "eslint": "^7.20.0",
+    "eslint-plugin-prettier": "^3.3.1",
+    "husky": "^5.1.0",
     "lemmy-js-client": "0.9.9",
     "lint-staged": "^10.5.4",
-    "mini-css-extract-plugin": "^1.3.6",
+    "mini-css-extract-plugin": "^1.3.8",
     "node-fetch": "^2.6.1",
     "node-sass": "^5.0.0",
     "prettier": "^2.2.1",
     "rimraf": "^3.0.2",
     "run-node-webpack-plugin": "^1.3.0",
-    "sass-loader": "^11.0.0",
+    "sass-loader": "^11.0.1",
     "sortpack": "^2.2.0",
     "style-loader": "^2.0.0",
-    "terser": "^5.6.0-beta",
-    "typescript": "^4.1.3",
-    "webpack": "5.21.2",
+    "terser": "^5.6.0",
+    "typescript": "^4.1.5",
+    "webpack": "5.23.0",
     "webpack-cli": "^4.5.0",
     "webpack-dev-server": "3.11.2",
     "webpack-node-externals": "^2.5.2"
@@ -88,11 +89,6 @@
     "node": ">=8.9.0"
   },
   "engineStrict": true,
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
-  },
   "lint-staged": {
     "*.{ts,tsx,js}": [
       "prettier --write",
