@@ -247,7 +247,7 @@ export class User extends Component<any, UserState> {
     };
   }
 
-  componentDidUpdate(lastProps: any, _lastState: UserState, _snapshot: any) {
+  componentDidUpdate(lastProps: any) {
     // Necessary if you are on a post and you click another post (same route)
     if (
       lastProps.location.pathname.split('/')[2] !==
